chore: improve file org and some cleanup
