+++ safenode/src/protocol/types/mod.rs
@@ -18,5 +18,7 @@ pub mod dbc_genesis;
 pub mod error;
 /// Register type.
 pub mod register;
+/// Token transfers.
+pub mod transfers;
 /// A wallet for network tokens.
 pub mod wallet;
